refactor(harness): collapse coding mode hosts into shared layers
