+++ docs/internals/architecture/harness/coding-to-harness-migration-inventory.md
@@ -87,7 +87,7 @@ entries.
 | `coding.bootstrap` | Keep product | Product assembly. It binds Coding's persisted Product-only capability profile to Harness resource/skill activation, prompt composition, and tool/command pack ordering; it may call harness engines but should not move. |
 | `coding.runtime` | Product adapter | Generic binding leases, runtime contexts, current-session transitions, serialized operation phases, uncommitted-candidate rollback, replacement callback ordering, exclusive import staging, navigation abort scopes, and coalesced scheduling live in Harness. `harness.session.SessionLifecycleRuntime` owns the common active-session new/restore/fork/import/replacement/disposal flow and default fork semantics; `harness.session.AgentTranscriptSessionRuntime` adds the standard Agent transcript directory/index and current-reference facade. Coding keeps composition, cwd/session-file resolution and acceptance policy, its `before` user-message fork interpretation, extension event projection, diagnostics codes, transcript semantics, package operations, and index content. |
 | `coding.ui` | Never harness | Product-owned TUI adapter and screen/controller state. Shared terminal primitives belong in `loushang.tui`, not harness. |
-| `coding.mode` | Product adapter | Print/RPC stay Coding-owned for command tables and legacy display contracts, but their JSON paths subscribe to the common RuntimeEvent stream and consume Coding-created RuntimeEventViews. `channel.JsonlCommandHost` owns the injected strict-JSON input loop, command validation, and dispatch-error isolation; `channel.RemoteUiContext` owns generic headless dialog correlation, timeout, and snapshot mechanics. Coding maps both into its established RPC responses, extension UI request type, tool enrichment, and Product event projection. The adapters retain explicit transport projection for known dataclasses, paths, mappings, lists, and tuples while rejecting cycles, sets, arbitrary objects, `__dict__` discovery, non-finite floats, and `repr()` fallback. |
+| `coding.mode` | Product adapter | `harness.host.rpc` owns the shared JSONL RPC host, standard session-operation handlers, state/model/diagnostic dispatch, task draining, and strict transport projection. `harnesstui.conversation.plain_mode` owns the plain/JSON output host and state observation. `channel` owns framing, correlation, cancellation, and delivery. Coding retains only the Work/domain operation binding plus injected event and diagnostic projections; no Pi aliases are retained. |
 | `coding.cli` | Keep product | Product CLI. It may expose harness-backed behavior but remains coding-owned. |
 | `coding.message` | Migrated and removed | `harness.conversation` owns the neutral envelope, repository, replay, and opaque-record behavior. The optional `harness.agent_transcript` profile owns standard Agent transcript payloads, codecs, state/context projection, the pure record factory, idempotent application-message commit, and an explicit Session v3 external importer. Native Product load accepts only the current format. Coding keeps only product presentation and orchestration policy. |
 | `coding.store` | Product adapter / compatibility shim | `ConversationStore`, revision/CAS semantics, Memory/File backends, and the open Agent transcript service live in Harness. The current Native Agent transcript codec, journal policy, lock, file layout, discovery, `FileConversationStore` assembly, standard session-facing commit/label/context operations, catalog summary/query/index/tree read model, lifecycle mechanics, and `AgentTranscriptSessionFactory` header/create/load/recent-resume/fork orchestration now live in `harness.agent_transcript`. Coding's `ProductRuntimePlan` selects file/memory stores and transcript profiles; its `SessionManager` facade retains root/persist decisions, runtime/capability profile binding and compatibility validation, Product-only index fields and retention, CLI/UI policy, and compatibility names. Database/Redis providers and journal-offset projection checkpoints remain deferred. |
