+++ src/loushang/harnesstui/conversation/__init__.py
@@ -1,2 +1,1 @@
 """Shared terminal interaction for projected conversations."""
-"""Reusable conversation UI primitives."""
